fix: filter depth data by visible intervals in soil-id export lookup
The soil-id algorithm call in fetch_soil_id() was passing all
depth-dependent data rows regardless of the active depth interval
preset. When a user switched presets (e.g., NRCS to BLM), old
measurements remained in the DB and were sent alongside new ones,
causing overlapping intervals that crashed pd.concat in the
soil-id-algorithm with InvalidIndexError.

Now reuses the same get_visible_intervals() logic used by CSV/JSON
export to filter measurements, respecting project-level preset
overrides. Extracted shared depth interval helpers into depth_helpers.py
to break the circular import between fetch_data and transformers.
